requirements.txt
pandas
plotly
umap-learn
rank-bm25
langchain
langchain-classic
langchain-community
langchain-ollama
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
beautifulsoup4
requests
streamlit
watchdog
ragas
datasets
python-dotenv
tiktoken
einops